% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StellarObjs.R
\docType{data}
\name{StellarObjs}
\alias{StellarObjs}
\title{Classification of Stars, Galaxies and Quasars from the Sloan Digital Sky Survey (SDSS)}
\format{
a data frame with 99,999 observations
\describe{
  \item{obj_ID}{The unique value that identifies the object in the image catalog used by CAS.}
  \item{right_ascension_angle}{Right Ascension Angle at J2000 epoch.}
  \item{declination_angle}{Declinatino angle at J2000 epoch.}
  \item{ultraviolet_filter}{Ultraviolet filter in the photometric system.}
  \item{green_filter}{Green filter in the photometric system.}
  \item{red_filter}{Red filter in the photometric system.}
  \item{near_infrared_filter}{Near Infrared filter in the photometric system.}
  \item{infrared filter}{Infrared filter in the photometric system.}
  \item{class}{object class: GALAXY, STAR or quasar (QSO) object.}
  \item{redshift}{redshift value based on the increase in wavelength}
  \item{date}{Date when a given piece of SDSS data was taken}
}
}
\source{
\url{https://www.kaggle.com/datasets/fedesoriano/stellar-classification-dataset-sdss17}
}
\usage{
StellarObjs
}
\description{
Classification of Stars, Galaxies and Quasars from the Sloan Digital Sky Survey (SDSS)
}
\keyword{datasets}
